LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-focus-previous-outside-expected.txt

button 1 button 2

PASS Focus should not be restored if the currently focused element is not inside the dialog.
PASS Focus restore should not occur when the focused element is in a shadowroot outside of the dialog.
PASS Focus restore should occur when the focused element is in a shadowroot inside the dialog.
FAIL Focus restore should occur when the focused element is slotted into a dialog. assert_equals: expected Element node <button id="b2">button 2</button> but got Element node <button id="host2button">button</button>
